@Embedded 사용
