+++ src/main/java/subway/service/LineService.java
@@ -47,7 +47,7 @@ public class LineService {
 
         line = lineRepository.save(line);
 
-        line.addSection(Section.builder()
+        line.getSections().addSection(Section.builder()
                 .upStation(upStation)
                 .downStation(downStation)
                 .distance(lineCreateRequest.getDistance())
@@ -98,7 +98,7 @@ public class LineService {
     @Transactional
     public void deleteLineSection(Long lineId, Long stationId) {
         Line line = getLineById(lineId);
-        line.deleteSection(stationId);
+        line.getSections().deleteSection(stationId);
     }
 
     public LineResponse findLine(Long id) {
@@ -112,7 +112,7 @@ public class LineService {
     }
 
     private static void validDownStation(Long downStationId, Line line) {
-        List<Long> registeredStationIds = line.getStationIds();
+        List<Long> registeredStationIds = line.getSections().getStationIds();
 
         if (registeredStationIds.contains(downStationId)) {
             throw new CheckDuplicateStationException("이미 해당노선에 등록되어있는 역은 새로운 구간의 하행역이 될 수 없습니다.");
@@ -120,7 +120,7 @@ public class LineService {
     }
 
     private static void validUpStation(Long upStationId, Line line) {
-        Long downStationId = line.findDownStationId();
+        Long downStationId = line.getSections().findDownStationId();
 
         if (downStationId != upStationId) {
             throw new InvalidUpStationException("새로운 구간의 상행역은 해당 노선에 등록되어 있는 하행 종점역이어야 합니다.");
